split image generation provider implementation
